+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugs Fixed
 
 ### Other Changes
+- Added `typing-extensions>=4.0.1` as a dependency.
 
 ## 12.10.0 (2022-10-11)